+++ packages/fuel-indexer-lib/Cargo.toml
@@ -11,10 +11,11 @@ description = "Fuel Indexer Library"
 
 [dependencies]
 anyhow = "1.0"
+bincode = { workspace = true }
 clap = { features = ["cargo", "derive", "env"], workspace = true }
 fuel-indexer-types = { workspace = true }
 http = { version = "0.2", default-features = false }
-serde = { features = ["derive"], workspace = true }
+serde = { workspace = true }
 serde_json = { workspace = true }
 serde_yaml = "0.8"
 sha2 = "0.9"
