+++ packages/fuel-indexer-schema/Cargo.toml
@@ -11,7 +11,6 @@ description = "Fuel Indexer Schema"
 
 [dependencies]
 async-graphql-parser = "5.0"
-bincode = { workspace = true }
 fuel-indexer-database = { workspace = true, optional = true }
 fuel-indexer-database-types = { workspace = true }
 fuel-indexer-types = { workspace = true }
